Combine impl files
